requirements/base.txt
click==8.1.3  # https://github.com/pallets/click/
httpx==0.23.3  # https://github.com/encode/httpx

# FastAPI
# ------------------------------------------------------------------------------
fastapi==0.105.0  # https://fastapi.tiangolo.com/
uvicorn[standard]==0.25.0  # https://github.com/encode/uvicorn
orjson==3.9.10  # https://github.com/ijl/orjson
jinja2==3.1.2  # https://github.com/pallets/jinja

fastapi-users[sqlalchemy,oauth]==12.1.2
httpx-oauth==0.13.1
aiosqlite==0.19.0

sqlalchemy==2.0.23
